Reorganize unit tests in liveblocks-redux
